Feat: Apply SearchBar onClick event
