+++ app/explore/_styles/searchHistories.module.scss
@@ -1,0 +1,73 @@
+.searchFnc {
+  width: 100%;
+  max-width: 430px;
+  .recordHeader {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    background-color: #fff;
+    font-size: 12px;
+    padding: 12px 16px;
+    color: #666;
+    font-weight: var(--fw-regular);
+
+    > button {
+      cursor: pointer;
+      font-size: 12px;
+      color: #666;
+    }
+  }
+}
+
+.searchRecordContainer {
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+
+  .recordSwitch {
+    display: flex;
+    align-items: center;
+    gap: 3px;
+    background-color: #fff;
+    font-size: 12px;
+  }
+
+  .searchRecords {
+    display: flex;
+    flex-direction: column;
+    font-size: 13px;
+    background-color: #fff;
+    padding-bottom: 85px;
+
+    &:first-child {
+      border: none;
+    }
+
+    .record {
+      display: flex;
+      justify-content: space-between;
+      padding: 0 16px;
+      gap: 15px;
+      border-top: 1px solid #f6f6f6;
+
+      > span {
+        cursor: pointer;
+        flex: 1;
+        padding: 17px 0;
+      }
+
+      & img {
+        cursor: pointer;
+      }
+    }
+  }
+  .noRecord {
+    position: absolute;
+    top: 45dvh;
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    font-size: 14px;
+    color: #666;
+  }
+}
