+++ app/search/_styles/search.module.scss
@@ -8,6 +8,7 @@
 }
 
 .postContainer {
+  flex: 1;
   padding: 0 16px 100px 16px;
   display: flex;
   flex-direction: column;
