+++ app/search/_styles/searchBar.module.scss
@@ -1,9 +1,11 @@
-.exploreBar {
+.searchBar {
+  position: relative;
   display: flex;
   justify-content: space-between;
   align-items: center;
   padding: 4px 16px;
   width: 100%;
+  max-width: 430px;
   background-color: #fff;
 
   > .searchInput {
@@ -30,5 +32,20 @@
         color: transparent;
       }
     }
+  }
+}
+
+.historiesContainer {
+  z-index: 100;
+  position: absolute;
+  top: 42px;
+  width: 100%;
+  height: calc(100% - 85px - 42px);
+  background-color: #fff;
+  overflow-y: scroll;
+  -ms-overflow-style: none; /* 인터넷 익스플로러 */
+  scrollbar-width: none; /* 파이어폭스 */
+  &::-webkit-scrollbar {
+    display: none;
   }
 }
